x/lto/types/message_create_anchor_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"lto/testutil/sample"
)

func TestMsgCreateAnchor_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateAnchor
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateAnchor{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateAnchor{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
